Separate interfaces for open and closed ranges
Closed ranges are now represented with `IClosedRange`,
and open ranges with `IOpenRange`.

Previously, for the first, `IRange` was used.
And the second inherited from `IRange`.
This made it impossible to have extension methods
with identical names for both open and closed ranges
at the same time.
Also, it was possible to substitute open range
as a closed one.
This could lead to unintented results, e.g.
and operation on closed range would accept an opened one,
and ignore the "open start" and "open end" properties.